% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnewscale-package.R
\docType{package}
\name{ggnewscale-package}
\alias{ggnewscale}
\alias{ggnewscale-package}
\title{ggnewscale: Multiple Fill and Color Scales in 'ggplot2'}
\description{
Use multiple fill and color scales in 'ggplot2'.
}
\section{Usage}{


To have more than one colour or fill scale in a plot using ggplot2, you add a
new scale slot with `new_scale()` and the following geoms will use a
different scale than the previous ones.
}

\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (0000-0002-7742-9230)

}
\keyword{internal}
