% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-aes.R
\name{+.ggplot_rename_next}
\alias{+.ggplot_rename_next}
\title{Add method to plots that have new scales}
\usage{
\method{+}{ggplot_rename_next}(e1, e2)
}
\arguments{
\item{e1}{PLot}

\item{e2}{element to add to the plot (layer, scale, guides, list of layers).}
}
\description{
Add method to plots that have new scales
}
