% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnewscale-package.R
\docType{package}
\name{ggnewscale-package}
\alias{ggnewscale}
\alias{ggnewscale-package}
\title{ggnewscale: Multiple Fill and Colour Scales in 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Use multiple fill and colour scales in 'ggplot2'.
}
\section{Usage}{


To have more than one colour or fill scale in a plot using ggplot2, you add a
new scale slot with \link{new_scale} and the following geoms will use a
different scale than the previous ones.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://eliocamp.github.io/ggnewscale/}
  \item Report bugs at \url{https://github.com/eliocamp/ggnewscale/issues}
}

}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
